+++ R/deming_power.R
@@ -195,6 +195,9 @@ deming_power_sim <- function(n_sims = 1000,
 #' @title Determine Required Sample Size for Deming Regression
 #'
 #' @description
+#'
+#' #' `r lifecycle::badge('experimental')`
+#'
 #' Automatically determines the minimum sample size needed to achieve target statistical
 #' power for detecting specified bias in method comparison studies using Deming regression.
 #'